Split send and receive events, and refresh automatic
